Moved logic from properties to propertyTable
